+++ src/main/java/org/ecommerce/services/impl/ProductServiceImpl.java
@@ -2,7 +2,6 @@ package org.ecommerce.services.impl;
 
 import org.ecommerce.models.Product;
 import org.ecommerce.services.ProductService;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
